// avenix-style-hub-main/src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.5"
  }
  public: {
    Tables: {
      advertisements: {
        Row: {
          created_at: string
          description: string | null
          id: string
          image_url: string | null
          is_active: boolean
          title: string | null
          updated_at: string
          video_url: string | null
        }
        Insert: {
          created_at?: string
          description?: string | null
          id?: string
          image_url?: string | null
          is_active?: boolean
          title?: string | null
          updated_at?: string
          video_url?: string | null
        }
        Update: {
          created_at?: string
          description?: string | null
          id?: string
          image_url?: string | null
          is_active?: boolean
          title?: string | null
          updated_at?: string
          video_url?: string | null
        }
        Relationships: []
      }
// ...
      order_items: {
        Row: {
          color: string | null
          created_at: string
          id: string
          order_id: string
          price: number
          product_id: string
          quantity: number
          size: string | null
        }
        Insert: {
          color?: string | null
          created_at?: string
          id?: string
          order_id: string
          price: number
          product_id: string
          quantity?: number
          size?: string | null
        }
        Update: {
          color?: string | null
          created_at?: string
          id?: string
          order_id?: string
          price?: number
          product_id?: string
          quantity?: number
          size?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "order_items_order_id_fkey"
            columns: ["order_id"]
            isOneToOne: false
            referencedRelation: "orders"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "order_items_product_id_fkey"
            columns: ["product_id"]
            isOneToOne: false
            referencedRelation: "products"
            referencedColumns: ["id"]
          },
        ]
      }
      orders: {
        Row: {
          address: Json
          created_at: string
          id: string
          order_status: string
          payment_method: string
          payment_status: string
          total_price: number
          transaction_id: string | null
          updated_at: string
          upi_app: string | null
          user_id: string
        }
        Insert: {
          address?: Json
          created_at?: string
          id?: string
          order_status?: string
          payment_method?: string
          payment_status?: string
          total_price: number
          transaction_id?: string | null
          updated_at?: string
          upi_app?: string | null
          user_id: string
        }
        Update: {
          address?: Json
          created_at?: string
          id?: string
          order_status?: string
          payment_method?: string
          payment_status?: string
          total_price?: number
          transaction_id?: string | null
          updated_at?: string
          upi_app?: string | null
          user_id?: string
        }
        Relationships: []
      }
// ...
      products: {
        Row: {
          category: string
          colors: Json
          created_at: string
          description: string | null
          id: string
          image: string
          images: string[]
          in_stock: boolean
          name: string
          original_price: number | null
          price: number
          rating: number
          reviews: number
          sizes: string[]
          stock_quantity: number
          tag: string | null
          updated_at: string
        }
        Insert: {
          category: string
          colors?: Json
          created_at?: string
          description?: string | null
          id?: string
          image: string
          images?: string[]
          in_stock?: boolean
          name: string
          original_price?: number | null
          price: number
          rating?: number
          reviews?: number
          sizes?: string[]
          stock_quantity?: number
          tag?: string | null
          updated_at?: string
        }
        Update: {
          category?: string
          colors?: Json
          created_at?: string
          description?: string | null
          id?: string
          image?: string
          images?: string[]
          in_stock?: boolean
          name?: string
          original_price?: number | null
          price?: number
          rating?: number
          reviews?: number
          sizes?: string[]
          stock_quantity?: number
          tag?: string | null
          updated_at?: string
        }
        Relationships: []
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      has_role: {
        Args: {
          _role: Database["public"]["Enums"]["app_role"]
          _user_id: string
        }
        Returns: boolean
      }
    }
    Enums: {
      app_role: "admin" | "user"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type TablesInsert<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Insert: infer I
    }
    ? I
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Insert: infer I
      }
      ? I
      : never
    : never

export type TablesUpdate<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Update: infer U
    }
    ? U
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Update: infer U
      }
      ? U
      : never
    : never